WIP normalization changes

+++ src/jef/GrayImage.java
@@ -5,22 +5,22 @@ public class GrayImage {
 	public static final int IMAGE_WIDTH = 220;
 	public static final int IMAGE_HEIGHT = 220;
 	
-	private int[] gImage = null;
+	private double[] gImage = null;
 	private int width = 0;
 	private int height = 0;
 	
-	public GrayImage(int[] imageData, int width, int height) {
+	public GrayImage(double[] imageData, int width, int height) {
 		gImage = imageData;
 		this.width = width;
 		this.height = height;
 	}
 	
-	public GrayImage(int[][] imageData) {
+	public GrayImage(double[][] imageData) {
 		if (imageData == null || imageData[0] == null) {
 			throw new IllegalArgumentException("GrayImage(): imageData is null in constructor");
 		}
 		int sublength = imageData[0].length;
-		gImage = new int[imageData.length * sublength];
+		gImage = new double[imageData.length * sublength];
 		for (int i = 0; i < imageData.length; ++i) {
 			if (imageData[i] == null || imageData[i].length != sublength) {
 				throw new IllegalArgumentException("GrayImage(): all sub-arrays must be non-null and of equal size");
@@ -33,7 +33,7 @@ public class GrayImage {
 		width = imageData[0].length;
 	}
 	
-	public int[] getImage() {
+	public double[] getImage() {
 		return gImage;
 	}
